Improve logrotate defaults to avoid endlessly growing log files.
I'd set up the default to move the old log file and create a new one in
its place. Stupidly, I'd forgotten that unless you tell the process
doing the logging to reopen the filename then it's just going to keep on
appending to the same file.

The *default* is copytruncate - copy the old log file and then truncate
it in place. That's not ideal but it doesn't need any cooperation from
the process doing the logging. See 'man logrotate' for details.